Move TEAM_VIDEO_PROJECTS_QUERY to ScrollableTableWithMenu
